fix: Solve comments from first PR review
