+++ mapping_suite_sdk/mapping_package_v3/adapters/metadata_serialiser.py
@@ -1,18 +1,19 @@
 from pathlib import Path
 
 from mapping_suite_sdk.core.adapters.serialiser import MappingPackageAssetSerialiser
-from mapping_suite_sdk.core.models.pydantic import fields
 from mapping_suite_sdk.mapping_package_v3.models.mapping_package_v3_metadata import MappingPackageV3Metadata
+from mapping_suite_sdk.mapping_package_v3.models.mapping_package_v3_metadata_jsonld import \
+    MappingPackageV3MetadataJSONLD
 
 
 class MappingPackageV3MetadataSerialiser(MappingPackageAssetSerialiser):
     """Serialiser for mapping package metadata."""
 
-    def serialise(self, package_folder_path: Path, asset: MappingPackageV3Metadata) -> None:
+    def serialise(self, package_folder_path: Path, asset: MappingPackageV3MetadataJSONLD) -> None:
         metadata_path = package_folder_path / asset.path
         metadata_path.parent.mkdir(parents=True, exist_ok=True)
-        metadata_path.write_text(asset.model_dump_json(by_alias=True,
-                                                       exclude={fields(MappingPackageV3Metadata).path},
-                                                       # For V3 (Unified)
-                                                       exclude_none=True,
-                                                       indent=4))
+        metadata = MappingPackageV3Metadata.model_construct(**asset.model_dump())
+        metadata_path.write_text(metadata.model_dump_json(by_alias=True,
+                                                          # For V3 (Unified)
+                                                          exclude_none=True,
+                                                          indent=4))
