+++ pyproject.toml
@@ -39,5 +39,5 @@ pytest-cov = "^6.0.0"
 coverage = "^7.6.12"
 pylint = "^3.3.4"
 mongomock = "^4.3.0"
-linkml = "^1.9.4"
+linkml = ">=1.9.2,<2.0.0"
 ruff = "^0.14.0"
