+++ resources/schema/mapping_package_v3/models/mapping_package_v3_metadata.yaml
@@ -1,5 +1,5 @@
-id: MappingPackageV3LinkMLMetadata
-name: MappingPackageV3LinkMLMetadata
+id: MappingPackageV3Metadata
+name: MappingPackageV3Metadata
 description: Schema for a data transformation rules package metadata
 prefixes:
   linkml: https://w3id.org/linkml/
@@ -10,7 +10,7 @@ imports:
   - linkml:types
 
 classes:
-  MappingPackageV3LinkMLMetadata:
+  MappingPackageV3Metadata:
     description: Metadata for a data transformation rules package
     slots:
       - id
@@ -96,13 +96,13 @@ slots:
     description: List of applicable document type identifiers
     multivalued: true
     examples:
-      - value: ["1", "24", "E1"]
+      - value: [ "1", "24", "E1" ]
 
   document_version_list:
     description: List of supported document versions
     multivalued: true
     examples:
-      - value: ["1.3", "1.7", "1.11"]
+      - value: [ "1.3", "1.7", "1.11" ]
 
   document_version_range:
     description: Range of (minimum and maximum) supported document versions
@@ -116,23 +116,23 @@ slots:
     description: Earliest applicable datetime for the mapping
     range: datetime
     examples:
-      - value: ["2014-01-01:T00:00:00Z"]
+      - value: [ "2014-01-01:T00:00:00Z" ]
 
   end:
     description: Latest applicable datetime for the mapping
     range: datetime
     examples:
-      - value: ["2014-01-30:T23:59:59Z"]
+      - value: [ "2014-01-30:T23:59:59Z" ]
 
   min:
     description: Minimum supported document version
     examples:
-      - value: ["R2.0.9.S01.E01"]
+      - value: [ "R2.0.9.S01.E01" ]
 
   max:
     description: Maximum supported document version
     examples:
-      - value: ["R2.0.9.S05.E01"]
+      - value: [ "R2.0.9.S05.E01" ]
 
   mapping_suite_hash_digest:
     required: true
