+++ resources/templates/base_model.py.jinja
